Dashboard: Soften edit links.
Change to anchor tags. To get them to appear/behave as hyperlinks,
set an href of #. To avoid browser 'jumping to the top' when clicking
a # link, set onclick to return false.

This is a temporary hack until we refactor the layout and behavior of
these cards.

Conduct the same for the list view.