Activity: Remove 'ongoing' attribute.
Users wish to specify activities in terms of start date to end date, and start time and end time.
In the case of one-time events, the start and end dates will be the same. There is need to
additionally distinguish multi-day/ongoing vs single-day/'one-time' activities.

Remove the field.

References #49.